+++ crates/live_kit_server/build.rs
@@ -1,0 +1,5 @@
+fn main() {
+    prost_build::Config::new()
+        .compile_protos(&["protocol/livekit_room.proto"], &["protocol"])
+        .unwrap();
+}
