+++ crates/live_kit_server/src/proto.rs
@@ -1,0 +1,1 @@
+include!(concat!(env!("OUT_DIR"), "/livekit.rs"));
